wl12xx: add sysfs control over disable SGI
The FW by default works with Short GI whenever possible.
This is causing trouble with some sniffers used in certification
labs. An option to disable SGI was added to the FW through
a generic command API. Add support for this new API
and enable control over disabling SGI through sysfs.
debugfs cannot be used here as the certification tests are being
run on production devices.
Note that we still publish support for SGI 20 in either STA/AP
however Short GI won't be used effectively when disable_sgi is 1.
This solution is good enough according to the certification folks.

(The generic cmd part is based on Luca's dropbcast patch
which was dropped)

Signed-off-by: Eyal Shapira <eyal@wizery.com>