wl12xx: clear RX filters when adding AP/GO interface
In AP/GO we don't want RX filters configured by upper levels
to be in effect. AP/GO has to be able to get all traffic
including broadcasts (for DHCP for example).
Once the last AP is removed we restore the filters.
Also fix a minor bug in some error paths in remove interface
where we might not go back to ELP sleep.

Signed-off-by: Eyal Shapira <eyal@wizery.com>