mac80211: fix sync of rx filters with driver and cfg80211
Avoid changing the local saved rx filters in case
the driver fails to set them. As the driver clears
the filters whenever the device is stopped we need
to configure them back when turning on the device to
avoid being out of sync with cfg80211 which reports
back to userspace what's configured.
Also fix missing locking in ieee80211_set_rx_filters.

Signed-off-by: Eyal Shapira <eyal@wizery.com>